requirements.txt
potassium
torch
transformers
eva-decord
salesforce-lavis
Pillow